requirements.txt
flask==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
sqlalchemy==2.0.41
python-dotenv==1.1.0
werkzeug==3.1.3
jinja2==3.1.6
wtforms==3.2.1
email-validator==2.1.1 